move plugins to lib folder
