+++ lib/plugins/base-logger.js
@@ -1,7 +1,7 @@
 'use strict';
 
-var Verbalize = require('../');
-var utils = require('../lib/utils');
+var Verbalize = require('../../');
+var utils = require('../utils');
 
 module.exports = function(options) {
   var opts = utils.extend({defaultListener: true}, options);
