+++ lib/plugins/colors.js
