+++ lib/plugins/rainbow.js
