+++ lib/plugins/styles.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var utils = require('../lib/utils');
+var utils = require('../utils');
 
 module.exports = function(options) {
   return function() {
